# openapi.yaml
openapi: 3.1.0
info:
  title: DataForGood API
  version: "1.0"
  description: |
    Route handlers de `app/api`. Las acciones del panel del SuperUsuario
    (`/sistema`, `/usuarios`, `/supervisar`) son server actions y no aparecen aquí.

    **Sesión de usuario:** se maneja con la cookie `session_token` (httpOnly). Ejecuta
    primero `POST /api/auth/login`; el navegador (o Insomnia) guarda la cookie
    y la manda sola en las siguientes peticiones.

    **SuperUsuario:** ejecuta `POST /api/auth/root/token` con la credencial de
    SuperUsuario del entorno (`ROOT_USER_ID` y su contraseña), copia `data.token`,
    pulsa **Authorize** y pégalo en `rootBearer`. Vale 2 horas. Con él responden
    `GET /api/usuarios` (todos los usuarios) y `GET /api/usuarios/{id}` (cualquier id).
    El resto de las rutas son de la cuenta de un usuario y piden `session_token`.

    **CSRF:** las mutaciones con un header `Origin` de otro sitio responden
    403 `Origen no permitido` (`proxy.ts`).

    Todas las respuestas JSON son `{ data }`, `{ message, data }` o `{ error }`.

    Este archivo se mantiene a mano: si cambias una ruta, actualízalo en el mismo commit.
# /api/docs/spec agrega al inicio el servidor desde el que se abre la página.
servers:
  - url: http://localhost:3000/dataforgood
    description: Docker local
  - url: http://localhost:3000
    description: next dev
  - url: https://multimodal-ai-lab.cicese.mx/dataforgood
    description: Producción
...
security:
  - sessionCookie: []
paths:
...
  /api/usuarios/{id}/contrasena:
    parameters:
      - $ref: "#/components/parameters/Id"
    patch:
      tags: [Usuarios]
      summary: Cambiar la propia contraseña
      description: |
        Solo la cuenta propia (403 con otro id) y solo si tiene contraseña (400 si es de Google).
        `nueva` debe tener 8 caracteres o más (y como máximo 72 bytes), una mayúscula, un número y
        un carácter especial, y ser distinta de `actual`. Una `actual` equivocada suma un intento
        fallido al mismo contador del login: con 5 la cuenta queda bloqueada 15 minutos (423).
        Las sesiones abiertas se conservan. Se registra en audit_log como `usuario.contrasena_cambiar`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [actual, nueva]
              properties:
                actual: { type: string, format: password }
                nueva: { type: string, format: password }
      responses:
        "200":
          description: Contraseña cambiada
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: object, properties: { ok: { type: boolean } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "423":
          description: Demasiados intentos fallidos; la cuenta está bloqueada temporalmente
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

  # ---------------------------------------------------------------- Campañas
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session_token
      description: Se obtiene con `POST /api/auth/login`. El navegador la manda sola.
    pendingVerificationCookie:
      type: apiKey
      in: cookie
      name: pending_verification_id
      description: La dejan el registro o un login sin verificar.
    rootBearer:
      type: http
      scheme: bearer
      description: Token del SuperUsuario. Se obtiene con `POST /api/auth/root/token` y vale 2 horas.
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string }
  responses:
...
    BadRequest:
      description: Datos inválidos
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Unauthorized:
      description: No hay sesión (o la cuenta está suspendida o baneada)
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Forbidden:
      description: Sin permiso para esta acción, o `Origen no permitido`
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    NotFound:
      description: No existe
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
...